+++ BTrees/intvaluemacros.h
@@ -29,7 +29,13 @@
 #define COPY_VALUE_FROM_ARG(TARGET, ARG, STATUS)                  \
   if (INT_CHECK(ARG)) {                                         \
       long vcopy = INT_AS_LONG(ARG);                            \
-      if (PyErr_Occurred()) { (STATUS)=0; (TARGET)=0; }           \
+      if (PyErr_Occurred()) {                                     \
+        if (PyErr_ExceptionMatches(PyExc_OverflowError)) {        \
+            PyErr_Clear();                                        \
+            PyErr_SetString(PyExc_TypeError, "integer out of range"); \
+        }                                                         \
+        (STATUS)=0; (TARGET)=0;                                   \
+      }                                                           \
       else if ((int)vcopy != vcopy) {                                  \
         PyErr_SetString(PyExc_TypeError, "integer out of range"); \
         (STATUS)=0; (TARGET)=0;                                   \
@@ -62,7 +68,13 @@
 #define COPY_VALUE_FROM_ARG(TARGET, ARG, STATUS)                  \
   if (INT_CHECK(ARG)) {                                           \
       long vcopy = INT_AS_LONG(ARG);                              \
-      if (PyErr_Occurred()) { (STATUS)=0; (TARGET)=0; }           \
+      if (PyErr_Occurred()) {                                     \
+        if (PyErr_ExceptionMatches(PyExc_OverflowError)) {        \
+            PyErr_Clear();                                        \
+            PyErr_SetString(PyExc_TypeError, "integer out of range"); \
+        }                                                         \
+        (STATUS)=0; (TARGET)=0;                                   \
+      }                                                           \
       else if (vcopy < 0) {                                       \
         PyErr_SetString(PyExc_TypeError, "can't convert negative value to unsigned int"); \
         (STATUS)=0; (TARGET)=0;                                   \